save
continue moving formulas from `other.py`
