+++ tests/software_tests/translator/data_record_definitions/test_formula.py
@@ -17,6 +17,8 @@ from uds.translator.data_record_definitions.formula import (
     REPORT_TYPE_2020,
     RESERVED_BIT,
     get_conditional_event_type_record_09_2020,
+    get_data,
+    get_data_from_memory,
     get_did_2013,
     get_did_2020,
     get_did_data_2013,
@@ -34,6 +36,8 @@ from uds.translator.data_record_definitions.formula import (
     get_event_type_record_07_2020,
     get_event_type_record_09_2020,
     get_formula_for_raw_data_record_with_length,
+    get_max_number_of_block_length,
+    get_max_number_of_block_length_file_transfer,
     get_memory_size_and_memory_address,
 )
 
@@ -206,6 +210,25 @@ class TestFunctions:
                                                          unit="bytes")],
                                                    any_order=False)
 
+    # get_max_number_of_block_length
+
+    @pytest.mark.parametrize("length_format_identifier", [0x00, 0x0F])
+    def test_get_max_number_of_block_length__value_error(self, length_format_identifier):
+        with pytest.raises(ValueError):
+            get_max_number_of_block_length(length_format_identifier)
+
+    @pytest.mark.parametrize("bytes_number, reserved", [
+        (0x1, 0x0),
+        (0xD, 0x3),
+        (0xF, 0x0),
+    ])
+    def test_get_max_number_of_block_length(self, bytes_number, reserved):
+        assert (get_max_number_of_block_length((bytes_number << 4) + reserved)
+                == (self.mock_raw_data_record.return_value,))
+        self.mock_raw_data_record.assert_called_once_with(name="maxNumberOfBlockLength",
+                                                          length=8 * bytes_number,
+                                                          unit="bytes")
+
     # get_did_records_formula_2020
 
     @patch(f"{SCRIPT_LOCATION}.get_dids_2020")
@@ -229,6 +252,34 @@ class TestFunctions:
         assert formula(mock_did_count) == mock_get_dids_2013.return_value
         mock_get_dids_2013.assert_called_once_with(did_count=mock_did_count,
                                                    record_number=mock_record_number)
+
+    # get_data_from_memory
+
+    @pytest.mark.parametrize("address_and_length_format_identifier", [0x00, 0x01, 0xF0])
+    def test_get_data_from_memory__value_error(self, address_and_length_format_identifier):
+        with pytest.raises(ValueError):
+            get_data_from_memory(address_and_length_format_identifier)
+
+    @pytest.mark.parametrize("memory_address_length, memory_size_length", [
+        (0x1, 0x1),
+        (0xD, 0x3),
+        (0xF, 0xF),
+    ])
+    def test_get_data_from_memory(self, memory_address_length, memory_size_length):
+        assert (get_data_from_memory((memory_size_length << 4) + memory_address_length)
+                == (self.mock_raw_data_record.return_value,))
+        self.mock_raw_data_record.assert_has_calls([call(name="memoryAddress",
+                                                         length=8 * memory_address_length),
+                                                    call(name="memorySize",
+                                                         length=8 * memory_size_length,
+                                                         unit="bytes"),
+                                                    call(name="Data from Memory",
+                                                         children=(self.mock_raw_data_record.return_value,
+                                                                   self.mock_raw_data_record.return_value),
+                                                         length=8 * (memory_address_length + memory_size_length),
+                                                         min_occurrences=1,
+                                                         max_occurrences=None)],
+                                                   any_order=True)
 
     # get_did_data_mask_2020
 
@@ -285,6 +336,35 @@ class TestFunctions:
         with pytest.raises(ValueError):
             input_kwargs["formula"](incorrect_did)
         assert input_kwargs["formula"](some_defined_did) == (self.mock_raw_data_record.return_value, )
+
+    # get_max_number_of_block_length_file_transfer
+
+    def test_get_max_number_of_block_length_file_transfer__value_error(self):
+        with pytest.raises(ValueError):
+            get_max_number_of_block_length_file_transfer(0x00)
+
+    @pytest.mark.parametrize("bytes_number", [0x01, 0xFF])
+    def test_get_max_number_of_block_length_file_transfer(self, bytes_number):
+        assert (get_max_number_of_block_length_file_transfer(bytes_number)
+                == (self.mock_raw_data_record.return_value,))
+        self.mock_raw_data_record.assert_called_once_with(name="maxNumberOfBlockLength",
+                                                          length=8 * bytes_number,
+                                                          unit="bytes")
+
+
+    # get_data
+
+    def test_get_data__value_error(self):
+        with pytest.raises(ValueError):
+            get_data(0)
+
+    @pytest.mark.parametrize("memory_size_length", [1, 23])
+    def test_get_data(self, memory_size_length):
+        assert get_data(memory_size_length) == (self.mock_raw_data_record.return_value,)
+        self.mock_raw_data_record.assert_called_once_with(name="data",
+                                                          length=8,
+                                                          min_occurrences=memory_size_length,
+                                                          max_occurrences=memory_size_length,)
 
     # get_event_type_record_01
