+++ uds/translator/data_record_definitions/__init__.py
@@ -6,7 +6,11 @@ from .conditional import (
     CONDITIONAL_CERTIFICATE_SERVER,
     CONDITIONAL_CHALLENGE_CLIENT,
     CONDITIONAL_CHALLENGE_SERVER,
+    CONDITIONAL_DATA,
+    CONDITIONAL_DATA_FROM_MEMORY,
     CONDITIONAL_EVENT_TYPE_RECORD_09_2020,
+    CONDITIONAL_MAX_NUMBER_OF_BLOCK_LENGTH,
+    CONDITIONAL_MAX_NUMBER_OF_BLOCK_LENGTH_FILE_TRANSFER,
     CONDITIONAL_MEMORY_ADDRESS_AND_SIZE,
     CONDITIONAL_OPTIONAL_ADDITIONAL_PARAMETER,
     CONDITIONAL_OPTIONAL_CHALLENGE_CLIENT,
@@ -95,15 +99,11 @@ from .other import (
     CHALLENGE_SERVER_LENGTH,
     COMMUNICATION_CONFIGURATION,
     CONDITIONAL_COMMUNICATION_CONTROL_REQUEST,
-    CONDITIONAL_DATA,
-    CONDITIONAL_DATA_FROM_MEMORY,
     CONDITIONAL_DIR_INFO,
     CONDITIONAL_FILE_AND_PATH_NAME,
     CONDITIONAL_FILE_SIZES,
     CONDITIONAL_FILE_SIZES_OR_DIR_INFO,
     CONDITIONAL_LINK_CONTROL_REQUEST,
-    CONDITIONAL_MAX_NUMBER_OF_BLOCK_LENGTH,
-    CONDITIONAL_MAX_NUMBER_OF_BLOCK_LENGTH_FILE_TRANSFER,
     CONDITIONAL_POWER_DOWN_TIME,
     CONDITIONAL_SECURED_DATA_TRANSMISSION_REQUEST,
     CONDITIONAL_SECURED_DATA_TRANSMISSION_RESPONSE,
